move function for normalized implications to the shared folder
