+++ src/lib/commons/types.ts
@@ -101,12 +101,6 @@ export type ImplicationDisplay = Replace<
 	}
 >
 
-export type NormalizedImplication = {
-	id: string
-	assumptions: Set<string>
-	conclusion: string
-}
-
 export type SearchResults = {
 	contradiction: string[] | null
 	satisfied_properties: string[]
