+++ src/lib/server/fetchers/implications.ts
@@ -1,13 +1,8 @@
 import { query } from '$lib/server/db.catdat'
 import sql from 'sql-template-tag'
 import { error } from '@sveltejs/kit'
-import type {
-	ImplicationDB,
-	NormalizedImplication,
-	StructureType
-} from '$lib/commons/types'
+import type { ImplicationDB, StructureType } from '$lib/commons/types'
 import { display_implication } from '$lib/server/transforms'
-import { parse_json_set, parse_nested_json_set } from '$shared/utils'
 
 export function fetch_implications(type: StructureType) {
 	const { rows, err } = query<ImplicationDB>(sql`
@@ -30,61 +25,4 @@ export function fetch_implications(type: StructureType) {
 	const implications = rows.map(display_implication)
 
 	return { type, implications }
-}
-
-/**
- * List of normalized implications of a given type that have no mapped assumptions
- * (e.g. no source and no target assumptions in the case of functors),
- * i.e. those that are universally true. Only those are relevant
- * for the consistency check.
- */
-export function get_normalized_implications(type: StructureType) {
-	// TODO: remove duplication with deduction script
-
-	const { rows, err } = query<{
-		id: string
-		is_equivalence: 0 | 1
-		assumptions: string
-		mapped_assumptions: string
-		conclusions: string
-	}>(
-		sql`
-			SELECT
-				id,
-				is_equivalence,
-				assumptions,
-				conclusions,
-				mapped_assumptions
-			FROM implications_view
-			WHERE type = ${type}
-		`
-	)
-
-	if (err) return { implications: null, err }
-
-	const implications: NormalizedImplication[] = []
-
-	for (const impl of rows) {
-		const assumptions = parse_json_set<string>(impl.assumptions)
-		const conclusions = parse_json_set<string>(impl.conclusions)
-		const mapped_assumptions = parse_nested_json_set<string>(impl.mapped_assumptions)
-
-		if (Object.values(mapped_assumptions).some((list) => list?.size)) continue
-
-		for (const conclusion of conclusions) {
-			implications.push({ id: impl.id, assumptions, conclusion })
-		}
-
-		if (impl.is_equivalence) {
-			for (const assumption of assumptions) {
-				implications.push({
-					id: impl.id,
-					assumptions: conclusions,
-					conclusion: assumption
-				})
-			}
-		}
-	}
-
-	return { implications, err: null }
 }
